+++ src/components/Admin/ListUsers/ListUser.js
@@ -1,11 +1,34 @@
 import React, { Component } from 'react';
 import $ from 'jquery';
+// import Dialog from 'material-ui/Dialog';
 
 
-export default class ListUser extends React.Component {
+
+export default class ListUser extends Component {
 
     componentDidMount() {
-        this.fetchUsers();
+
+        let url;
+        url = "http://api.susi.ai/aaa/account-permissions.json";
+        $.ajax({
+            url: url,
+            dataType: 'jsonp',
+            jsonpCallback: 'py',
+            jsonp: 'callback',
+            crossDomain: true,
+            success: function (response) {
+                console.log(response.userRole)
+                if(response.userRole!=="admin") {
+                    console.log("Not an admin")
+                } else {
+                    this.fetchUsers();
+                    console.log("Admin")
+                }
+            }.bind(this),
+            error: function (errorThrown) {
+                console.log(errorThrown)
+            }
+        });
     }
 
     fetchUsers = () => {
